+++ Sources/Sentry/Public/SentryAppStartMeasurement.h
@@ -13,9 +13,6 @@ SENTRY_NO_INIT
 
 /**
  * Initializes SentryAppStartMeasurement with the given parameters.
- *
- * @param type The type of the app start. Either cold or warm.
- * @param duration The duration of the app start.
  */
 - (instancetype)initWithType:(SentryAppStartType)type
               appStartTimestamp:(NSDate *)appStartTimestamp
@@ -24,19 +21,29 @@ SENTRY_NO_INIT
     didFinishLaunchingTimestamp:(NSDate *)didFinishLaunchingTimestamp;
 
 /**
- * The type of the app start. Either cold or warm.
+ * The type of the app start.
  */
 @property (readonly, nonatomic, assign) SentryAppStartType type;
 
 /**
- * How long the app start toock.
+ * How long the app start took. From appStartTimestamp to when the SDK creates the
+ * AppStartMeasurement, which is done when the OS posts UIWindowDidBecomeVisibleNotification.
  */
 @property (readonly, nonatomic, assign) NSTimeInterval duration;
 
+/**
+ * The timestamp when the app started, which is is the proces start timestamp.
+ */
 @property (readonly, nonatomic, strong) NSDate *appStartTimestamp;
 
+/**
+ * When the runtime was initialized / when SentryAppStartTracker is added to the Objective-C runtime
+ */
 @property (readonly, nonatomic, strong) NSDate *runtimeInitTimestamp;
 
+/**
+ * When OS posts UIApplicationDidFinishLaunchingNotification.
+ */
 @property (readonly, nonatomic, strong) NSDate *didFinishLaunchingTimestamp;
 
 @end