+++ Sources/Sentry/SentryNetworkTracker.m
@@ -1,4 +1,5 @@
 #import "SentryNetworkTracker.h"
+#import "SentryBreadcrumb.h"
 #import "SentryHttpInterceptor+Private.h"
 #import "SentryHub+Private.h"
 #import "SentryLog.h"
@@ -142,18 +143,37 @@ SentryNetworkTracker ()
             OBJC_ASSOCIATION_RETAIN_NONATOMIC);
     }
 
+    SentryLevel breadcrumbLevel = sessionTask.error != nil ? kSentryLevelError : kSentryLevelInfo;
+    SentryBreadcrumb *breadcrumb = [[SentryBreadcrumb alloc] initWithLevel:breadcrumbLevel
+                                                                  category:@"http"];
+    breadcrumb.type = @"http";
+    NSMutableDictionary<NSString *, id> *breadcrumbData = [NSMutableDictionary new];
+    breadcrumbData[@"url"] = sessionTask.currentRequest.URL.absoluteString;
+    breadcrumbData[@"method"] = sessionTask.currentRequest.HTTPMethod;
+    breadcrumbData[@"request_body_size"] =
+        [NSNumber numberWithLongLong:sessionTask.countOfBytesSent];
+    breadcrumbData[@"response_body_size"] =
+        [NSNumber numberWithLongLong:sessionTask.countOfBytesReceived];
+
+    NSInteger responseStatusCode = [self urlResponseStatusCode:sessionTask.response];
+
+    if (responseStatusCode != -1) {
+        NSNumber *statusCode = [NSNumber numberWithInteger:responseStatusCode];
+        [netSpan setTagValue:[NSString stringWithFormat:@"%@", statusCode]
+                      forKey:@"http.status_code"];
+        breadcrumbData[@"status_code"] = statusCode;
+        breadcrumbData[@"reason"] =
+            [NSHTTPURLResponse localizedStringForStatusCode:responseStatusCode];
+    }
+
+    breadcrumb.data = breadcrumbData;
+    [SentrySDK addBreadcrumb:breadcrumb];
+
     if (netSpan == nil) {
         return;
     }
 
     [sessionTask removeObserver:self forKeyPath:NSStringFromSelector(@selector(state))];
-
-    NSInteger responseStatusCode = [self urlResponseStatusCode:sessionTask.response];
-
-    if (responseStatusCode != -1) {
-        [netSpan setTagValue:[NSString stringWithFormat:@"%li", (long)responseStatusCode]
-                      forKey:@"http.status_code"];
-    }
 
     [netSpan setDataValue:sessionTask.currentRequest.HTTPMethod forKey:@"method"];
     [netSpan setDataValue:sessionTask.currentRequest.URL.path forKey:@"url"];