+++ Tests/SentryTests/Integrations/TestDisplayLinkWrapper.swift
@@ -25,6 +25,28 @@ class TestDiplayLinkWrapper: SentryDisplayLinkWrapper {
         target = nil
         selector = nil
     }
+    
+    func givenFrames(_ slow: Int, _ frozen: Int, _ normal: Int) {
+        self.call()
+
+        // Slow frames
+        for _ in 0..<slow {
+            self.internalTimestamp += TestData.slowFrameThreshold + 0.001
+            self.call()
+        }
+
+        // Frozen frames
+        for _ in 0..<frozen {
+            self.internalTimestamp += TestData.frozenFrameThreshold + 0.001
+            self.call()
+        }
+
+        // Normal frames.
+        for _ in 0..<(normal - 1) {
+            self.internalTimestamp += TestData.slowFrameThreshold - 0.01
+            self.call()
+        }
+    }
 }
 
 #endif