+++ Tests/SentryTests/Performance/SentryTracer+Test.h
@@ -1,0 +1,12 @@
+
+#import "SentryTracer.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface SentryTracer (Test)
+
++ (void)resetAppStartMeasurmentRead;
+
+@end
+
+NS_ASSUME_NONNULL_END